docs: update component descriptions
feat: move navigation to separate component folder
