+++ src/components/header/Header.tsx
@@ -1,26 +1,10 @@
 /** @jsxImportSource theme-ui */
-import {
-	Children,
-	createElement,
-	forwardRef,
-	isValidElement,
-	useEffect,
-	useState,
-} from 'react';
-import { animateScroll } from 'react-scroll';
+import { Children, forwardRef, isValidElement } from 'react';
 import { ThemeUICSSObject } from 'theme-ui';
 import { classNames } from '@src/utils/classNames';
-import { SCREEN_MD } from '@src/constants/breakpoints';
 import { HeaderLogo } from '../logo';
-import {
-	Header,
-	Nav,
-	NavLinksContainer,
-	NavLogoContainer,
-	NavMobileButton,
-	NavMobileMenu,
-	NavMobileMenuItem,
-} from './Header.styled';
+import { PageNavigation } from '../navigation/Navigation';
+import { Header } from './Header.styled';
 
 export interface HeaderProps extends React.HTMLAttributes<HTMLDivElement> {
 	/**
@@ -49,85 +33,16 @@ export interface HeaderProps extends React.HTMLAttributes<HTMLDivElement> {
 	sx?: ThemeUICSSObject;
 }
 
-export const PageMobileNav = ({
-	blockLinks,
-}: {
-	blockLinks?: React.ReactNode[];
-}) => {
-	const [isMobileNavOpen, setIsMobileNavOpen] = useState(false);
-
-	useEffect(() => {
-		const handleResize = () => {
-			if (window.innerWidth > SCREEN_MD) {
-				setIsMobileNavOpen(false);
-			}
-		};
-
-		window.addEventListener('resize', handleResize);
-
-		return () => {
-			window.removeEventListener('resize', handleResize);
-		};
-	}, []);
-
-	return (
-		<>
-			<NavMobileButton
-				id="loom-header_nav-mobile-button"
-				onClick={() => setIsMobileNavOpen((prev) => !prev)}
-			>
-				<svg
-					width="24"
-					height="24"
-					viewBox="0 0 24 24"
-					fill="none"
-					xmlns="http://www.w3.org/2000/svg"
-				>
-					{isMobileNavOpen ? (
-						<path
-							d="M6 18L18 6M6 6L18 18"
-							stroke="currentColor"
-							strokeWidth="2"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-						/>
-					) : (
-						<path
-							d="M4 6H20M4 12H20M4 18H20"
-							stroke="currentColor"
-							strokeWidth="2"
-							strokeLinecap="round"
-							strokeLinejoin="round"
-						/>
-					)}
-				</svg>
-			</NavMobileButton>
-			<NavMobileMenu $isOpen={isMobileNavOpen}>
-				{blockLinks &&
-					blockLinks.map((link: React.ReactNode, index: number) => (
-						<NavMobileMenuItem
-							key={`mobile-menu-item-${index}`}
-							className="loom-header_nav-mobile-menu-item"
-						>
-							{createElement((link as React.ReactElement).type, {
-								...(link as React.ReactElement).props,
-								onClick: () => setIsMobileNavOpen(false),
-							})}
-						</NavMobileMenuItem>
-					))}
-			</NavMobileMenu>
-		</>
-	);
-};
-
 /**
- * TODO: Add description
+ * `PageHeader` component is a section at the top of the page. It can hold your site's logo and navigation links. The component supports block snapping and has an `sx` prop for custom styles. It can account for a fixed navigation bar if `fixedNav` prop is set to `true` in the `PageWrapper` component.
  *
  * @example
- * <Header>
+ * <PageHeader fixedNav={true}>
+ *   <HeaderLogo>MySite</HeaderLogo>
  *   <!-- your content here -->
- * </Header>
+ * </PageHeader>
  */
+
 export const PageHeader = forwardRef<HTMLDivElement, HeaderProps>(
 	(
 		{
@@ -161,27 +76,15 @@ export const PageHeader = forwardRef<HTMLDivElement, HeaderProps>(
 				className={classes}
 				$snap={snap}
 				$fixedNav={fixedNav}
-				{...rest}
 				ref={ref}
+				{...rest}
 			>
 				{!fixedNav && (
-					<Nav $fixed={fixedNav} id="loom-header_nav">
-						<NavLogoContainer
-							className="loom-header_nav-logo"
-							onClick={() => {
-								animateScroll.scrollToTop({
-									duration: 500,
-									containerId: 'loom-wrapper',
-								});
-							}} // TODO: Add duration prop
-						>
-							{logoElement}
-						</NavLogoContainer>
-						<NavLinksContainer className="loom-header_nav-links">
-							{blockLinks}
-						</NavLinksContainer>
-						<PageMobileNav blockLinks={blockLinks} />
-					</Nav>
+					<PageNavigation
+						blockLinks={blockLinks}
+						fixedNav={fixedNav}
+						logoElement={logoElement}
+					/>
 				)}
 				{childrenElements}
 			</Header>
